requirements.txt
blinker==1.6.2
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
filelock==3.12.2
Flask==2.3.2
grpcio==1.53.0
grpcio-tools==1.53.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mpmath==1.3.0
networkx==3.1
numpy==1.25.2
param==1.13.0
Pillow==10.0.0
protobuf==4.21.12
python-dotenv==1.0.0
requests==2.31.0
stability-sdk==0.8.4
sympy==1.12
torch==2.0.1
torchvision==0.15.2
typing_extensions==4.7.1
urllib3==2.0.4
Werkzeug==2.3.6
